requirements.txt
requests>=2.24.0
luigi[toml]>=3.0.2
xlrd==1.2.0
attrs>=20.2.0
pandas>=1.1.3
python-box>=5.1.1
python-dotenv
sqlalchemy
psycopg2-binary